LayoutTests/platform/gtk/accessibility/aria-combobox-expected.txt

option 1
option 2
This tests that the aria roles for combobox and aria-expanded work correctly in conjunction.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Role: AXRole: AXComboBox
PASS combobox.isExpanded is false
PASS combobox.isExpanded is false
Role: AXRole: AXListItem
PASS successfullyParsed is true

TEST COMPLETE
